Update MS_Text_Search_Purposym_And_Synonym_Thesauruses_Integration_T-SQL_SQL_Server_2019_IOC_v01.03.13_MAJOR.txt
A slight change to the query here. I have made sure that the full purposym table variable uses a table of fragments which doesn't contain any duplicates. This will make the subsequent synonym technique more effective. I have used a delete method rather than a group method, because it is more efficient to do this when cross joins and massive numbers of rows of data (fragments) are involved. I haven't tested the refactored code, so there may be one or two minor bugs that need fixing.