FINAL VERSION OF THE MICHAELIAN PROBABILITY ALGORITHM! (BUG ALERT))
Happy Christmas. This is an early present. Please be aware that the underlying SQL Server 2019 engine actually has at least 1 showstopper table variable joining feature bug (which manifest/s on 3 of the table variable transformations contained within). As a result I advise you firstly to definitely not deploy this algorithm to production, and secondly, to alert Microsoft to these terrible bugs... They could be corrupting thousands/millions of databases around the world as I write this!