"She ain't pretty, but she's good at what she does!... Testing was 'painfully shallow' and 'somewhat provisional'... Windows (10) operating systems only, please!"
Delivery took a month of half days worth of work. Half a day of unit testing was carried out. Integration, end-to-end, user acceptance, and chaos testing are deemed to be the responsibility of each Jedi Temple!

This version is a slight/cosmetic update (extra guidance in the auto-audit checking functionality).