"She ain't pretty, but she's good at what she does!... Testing was 'painfully shallow' and 'somewhat provisional'... Windows 10 Operating Systems 'only', please!"
Delivery took a month of half days worth of work. Half a day of unit testing was carried out. Integration, end-to-end, user-acceptance, and chaos testing are deemed to be the responsibility of each Jedi Temple!

This version massively improves speed of execution termination for one auto-audit checking case, where there is a lot of data in the main spreadsheet table.

You may need to right-click the file once downloaded, then click on the properties option, and then at the bottom of the pop-up settings box, in the Attributes section, tick-click the unblock option... (at least in Windows 10 Operating Systems)...