Obsidian Partitioned Number Sorter Release
I have not done any testing on this algorithm. Just wanted to get it out there, because number dataset sorting is such an ubiquitous and compute time expensive and energy-efficiency expensive algorithm... Especially in machine learning.

Please read the nota bene in PART_B before using in production, as there are certain conditions that must be met in order for massive performance gains to be unlocked by this algorithm, in comparison to current best-in-class approaches such as merge-sort etc_